Move spark-operator
