+++ cmd/operator/app/controller/root.go
