+++ cmd/operator/app/version/root.go
